change tasks to be objects instead of simple strings
The delete feature needs each task to have a unique id (as well as
the key required by react).

This requires us to change a task to be an object so that we can
hold multiple values for a task.

The task gets its unique id when it is created via the addTask
component and then this is used to filter it out when removing items
from the list.